requirements.txt
# Django 
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
# Authentication
django-cors-headers==4.3.1
djoser==2.2.2
PyJWT==2.8.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
# SQL database
psycopg2==2.9.9
# Images
pillow==10.3.0
# boto3==1.34.11
# botocore==1.34.11
# django-storages==1.14.2
# Web3
# eth-account==0.10.0
hdwallet==2.2.1
bitcoin-utils==0.6.6
# Fitures
python-slugify==8.0.1
django-environ==0.11.2
django-ses==3.5.2
python-dotenv==1.0.1